requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
Pillow==7.2.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
